Python/python-backend/requirements.txt
Flask
pandas
jsonify
psycopg2-binary
